Working on revision 2.
Diode is not strangely routed to ground any more, it is more sensible and direct.
Ground plane moved to the back so that it can print as a 2 sided board with the
silk screen on top (currently it is a one sided one with silk screen on the back
accidentally).